Add a `-w` option.
One problem with hot keys is that if you're still holding down, for
example, Control, then surprising things can happen when commands
execute (`xdotool type` is a great example, where you might expect "o"
to be typed but X will interpret it as Ctrl+O if you've still got
Control held down).

The `-w` option waits for all keys (not just the hotkeys!) to be
released before executing the command. I'm not entirely sure how to do
this "nicely" so we use a busy loop (with a 0.01s sleep) to monitor the
keyboard until all keys are released. A better mechanism is probably
possible, but this is surprisingly effective.