NoSymbol isn't a valid KeyCode.
Well, it sort of is, since it has a value of 0 -- but it's confusing to
see it used in this way. Since KeyCodes have (for reasons I don't
understand) a valid range of 8..255 (both ends inclusive), it seems we
can safely use 0 too -- but GrabKey uses 0 to represent `AnyKey`, so
it's safest to use a value in 1--7 (both ends inclusive).