Ignore the Mode Switch key and ignore modifiers with `-w`.
At least on my Framework, the "Mode Switch" key (which, I admit, I had
not previously heard of) is permanently set as "pressed" so `-w` blocks
indefinitely.

This commit first adds "Mode Switch" to the list of ignored keys. It
then explicitly ignores any modifiers when `-w` is set. There may be
consequences to this that I don't currently understand, but it Works For
Me.